idk claude code took the wheel when asked to refactor codebase
