+++ DEBUG_current_listening_track.json
@@ -1,0 +1,1 @@
+{"id":"","timestamp":1753141294178,"progress_ms":22811,"item":{"album":{"album_type":"single","total_tracks":1,"images":[{"url":"https://i.scdn.co/image/ab67616d0000b273e8517d37c3cc16895acdafdb","height":640,"width":640},{"url":"https://i.scdn.co/image/ab67616d00001e02e8517d37c3cc16895acdafdb","height":300,"width":300},{"url":"https://i.scdn.co/image/ab67616d00004851e8517d37c3cc16895acdafdb","height":64,"width":64}],"name":"we stayed up all night","release_date":"2025-02-07","release_date_precision":"day","type":"album","artists":[{"href":"https://api.spotify.com/v1/artists/788qKGMEh4hfYUTy8yANRC","id":"788qKGMEh4hfYUTy8yANRC","name":"Conrad.","type":"artist","uri":"spotify:artist:788qKGMEh4hfYUTy8yANRC"}]}}}
